[JDBC] Moved JDbc stuffs out of common plugin
