+++ .idea/compiler.xml
@@ -7,6 +7,7 @@
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
         <module name="protocol-common" />
+        <module name="protocol-common-jdbc" />
         <module name="protocol-mongo" />
         <module name="protocol-test" />
         <module name="jacoco" />
