+++ pom.xml
@@ -20,6 +20,7 @@
         <module>jacoco</module>
         <module>protocol-http</module>
         <module>sample-plugins</module>
+        <module>protocol-common-jdbc</module>
     </modules>
 
     <properties>
