+++ protocol-common-jdbc/pom.xml
@@ -1,13 +1,22 @@
-<project>
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-
     <parent>
         <groupId>org.kendar.protocol</groupId>
         <artifactId>protocol-master</artifactId>
         <version>${revision}</version>
     </parent>
 
-    <artifactId>protocol-mysql</artifactId>
+    <artifactId>protocol-common-jdbc</artifactId>
+
+    <properties>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    </properties>
+
     <dependencies>
         <dependency>
             <groupId>org.pf4j</groupId>
@@ -27,19 +36,10 @@
         </dependency>
         <dependency>
             <groupId>org.kendar.protocol</groupId>
-            <artifactId>protocol-common-jdbc</artifactId>
-            <version>${revision}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.mysql</groupId>
-            <artifactId>mysql-connector-j</artifactId>
-            <version>${mysql.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.kendar.protocol</groupId>
             <artifactId>protocol-test</artifactId>
             <version>${revision}</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
+
 </project>
