+++ protocol-common-jdbc/src/main/java/org/kendar/plugins/JdbcRecordingPlugin.java
