+++ protocol-common-jdbc/src/main/java/org/kendar/plugins/JdbcRewritePlugin.java
