+++ protocol-common-jdbc/src/main/java/org/kendar/sql/jdbc/DataTypeDescriptor.java
