+++ protocol-common-jdbc/src/main/java/org/kendar/sql/jdbc/ProxyMetadata.java
