+++ protocol-common-jdbc/src/main/java/org/kendar/sql/jdbc/SelectColumn.java
