+++ protocol-common-jdbc/src/main/java/org/kendar/sql/jdbc/SelectResult.java
