+++ protocol-common-jdbc/src/main/java/org/kendar/sql/jdbc/settings/JdbcProtocolSettings.java
