+++ protocol-common-jdbc/src/main/java/org/kendar/sql/parser/SqlStringParser.java
