+++ protocol-common-jdbc/src/main/java/org/kendar/sql/parser/SqlStringType.java
