+++ protocol-postgres/pom.xml
@@ -32,6 +32,11 @@
         </dependency>
         <dependency>
             <groupId>org.kendar.protocol</groupId>
+            <artifactId>protocol-common-jdbc</artifactId>
+            <version>${revision}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.kendar.protocol</groupId>
             <artifactId>protocol-test</artifactId>
             <version>${revision}</version>
             <scope>test</scope>
